+++ fatiando/gravmag/tesseroid.py
@@ -22,16 +22,16 @@ Gravity
 -------
 
 Forward modeling of gravitational fields is performed by functions:
-:func:`~fatiando.gravmag.prism.potential`,
-:func:`~fatiando.gravmag.prism.gx`,
-:func:`~fatiando.gravmag.prism.gy`,
-:func:`~fatiando.gravmag.prism.gz`,
-:func:`~fatiando.gravmag.prism.gxx`,
-:func:`~fatiando.gravmag.prism.gxy`,
-:func:`~fatiando.gravmag.prism.gxz`,
-:func:`~fatiando.gravmag.prism.gyy`,
-:func:`~fatiando.gravmag.prism.gyz`,
-:func:`~fatiando.gravmag.prism.gzz`
+:func:`~fatiando.gravmag.tesseroid.potential`,
+:func:`~fatiando.gravmag.tesseroid.gx`,
+:func:`~fatiando.gravmag.tesseroid.gy`,
+:func:`~fatiando.gravmag.tesseroid.gz`,
+:func:`~fatiando.gravmag.tesseroid.gxx`,
+:func:`~fatiando.gravmag.tesseroid.gxy`,
+:func:`~fatiando.gravmag.tesseroid.gxz`,
+:func:`~fatiando.gravmag.tesseroid.gyy`,
+:func:`~fatiando.gravmag.tesseroid.gyz`,
+:func:`~fatiando.gravmag.tesseroid.gzz`
 
 The gravitational fields are calculated using the formula of Grombein et al.
 (2013):
@@ -171,14 +171,21 @@ RATIO_GG = 8
 STACK_SIZE = 100
 
 
-def _check_input(lon, lat, height, model, ratio, njobs):
+def _check_input(lon, lat, height, model, ratio, njobs, pool):
     """
     Check if the inputs are as expected and generate the output array.
+
+    Returns:
+
+    * results : 1d-array, zero filled
+
     """
     assert lon.shape == lat.shape == height.shape, \
         "Input coordinate arrays must have same shape"
     assert ratio > 0, "Invalid ratio {}. Must be > 0.".format(ratio)
     assert njobs > 0, "Invalid number of jobs {}. Must be > 0.".format(njobs)
+    if njobs == 1:
+        assert pool is None, "njobs should be number of processes in the pool"
     result = np.zeros_like(lon)
     return result
 
@@ -245,7 +252,40 @@ def _check_tesseroid(tesseroid, dens):
     return density
 
 
-def _dispatcher(args):
+def _dispatcher(field, lon, lat, height, model, **kwargs):
+    """
+    Dispatch the computation of *field* to the appropriate function.
+
+    Returns:
+
+    * result : 1d-array
+
+    """
+    njobs = kwargs.get('njobs', 1)
+    pool = kwargs.get('pool', None)
+    engine = kwargs['engine']
+    dens = kwargs['dens']
+    ratio = kwargs['ratio']
+    result = _check_input(lon, lat, height, model, ratio, njobs, pool)
+    if njobs > 1 and pool is None:
+        pool = multiprocessing.Pool(njobs)
+        created_pool = True
+    else:
+        created_pool = False
+    if pool is None:
+        _forward_model([lon, lat, height, result, model, dens, ratio, engine,
+                        field])
+    else:
+        chunks = _split_arrays(arrays=[lon, lat, height, result],
+                               extra_args=[model, dens, ratio, engine, field],
+                               nparts=njobs)
+        result = np.hstack(pool.map(_forward_model, chunks))
+    if created_pool:
+        pool.close()
+    return result
+
+
+def _forward_model(args):
     """
     Run the computations on the model for a given list of arguments.
 
@@ -294,7 +334,7 @@ def _split_arrays(arrays, extra_args, nparts):
 
 
 def potential(lon, lat, height, model, dens=None, ratio=RATIO_V,
-              engine='default', njobs=1):
+              engine='default', njobs=1, pool=None):
     """
     Calculate the gravitational potential due to a tesseroid model.
 
@@ -326,6 +366,11 @@ def potential(lon, lat, height, model, dens=None, ratio=RATIO_V,
     * njobs : int
         Split the computation into *njobs* parts and run it in parallel using
         ``multiprocessing``. If ``njobs=1`` will run the computation in serial.
+    * pool : None or multiprocessing.Pool object
+        If not None, will use this pool to run the computation in parallel
+        instead of creating a new one. You must still specify *njobs* as the
+        number of processes in the pool. Use this to avoid spawning processes
+        on each call to this functions, which can have significant overhead.
 
     Returns:
 
@@ -333,24 +378,15 @@ def potential(lon, lat, height, model, dens=None, ratio=RATIO_V,
         The calculated field in SI units
 
     """
-    result = _check_input(lon, lat, height, model, ratio, njobs)
     field = 'potential'
-    if njobs == 1:
-        _dispatcher([lon, lat, height, result, model, dens, ratio, engine,
-                     field])
-    else:
-        chunks = _split_arrays(arrays=[lon, lat, height, result],
-                               extra_args=[model, dens, ratio, engine, field],
-                               nparts=njobs)
-        pool = multiprocessing.Pool(njobs)
-        result = np.hstack(pool.map(_dispatcher, chunks))
-        pool.close()
+    result = _dispatcher(field, lon, lat, height, model, dens=dens,
+                         ratio=ratio, engine=engine, njobs=njobs, pool=pool)
     result *= G
     return result
 
 
 def gx(lon, lat, height, model, dens=None, ratio=RATIO_G, engine='default',
-       njobs=1):
+       njobs=1, pool=None):
     """
     Calculate the North component of the gravitational attraction.
 
@@ -382,6 +418,11 @@ def gx(lon, lat, height, model, dens=None, ratio=RATIO_G, engine='default',
     * njobs : int
         Split the computation into *njobs* parts and run it in parallel using
         ``multiprocessing``. If ``njobs=1`` will run the computation in serial.
+    * pool : None or multiprocessing.Pool object
+        If not None, will use this pool to run the computation in parallel
+        instead of creating a new one. You must still specify *njobs* as the
+        number of processes in the pool. Use this to avoid spawning processes
+        on each call to this functions, which can have significant overhead.
 
     Returns:
 
@@ -389,24 +430,15 @@ def gx(lon, lat, height, model, dens=None, ratio=RATIO_G, engine='default',
         The calculated field in mGal
 
     """
-    result = _check_input(lon, lat, height, model, ratio, njobs)
     field = 'gx'
-    if njobs == 1:
-        _dispatcher([lon, lat, height, result, model, dens, ratio, engine,
-                     field])
-    else:
-        chunks = _split_arrays(arrays=[lon, lat, height, result],
-                               extra_args=[model, dens, ratio, engine, field],
-                               nparts=njobs)
-        pool = multiprocessing.Pool(njobs)
-        result = np.hstack(pool.map(_dispatcher, chunks))
-        pool.close()
+    result = _dispatcher(field, lon, lat, height, model, dens=dens,
+                         ratio=ratio, engine=engine, njobs=njobs, pool=pool)
     result *= SI2MGAL*G
     return result
 
 
 def gy(lon, lat, height, model, dens=None, ratio=RATIO_G, engine='default',
-       njobs=1):
+       njobs=1, pool=None):
     """
     Calculate the East component of the gravitational attraction.
 
@@ -438,6 +470,11 @@ def gy(lon, lat, height, model, dens=None, ratio=RATIO_G, engine='default',
     * njobs : int
         Split the computation into *njobs* parts and run it in parallel using
         ``multiprocessing``. If ``njobs=1`` will run the computation in serial.
+    * pool : None or multiprocessing.Pool object
+        If not None, will use this pool to run the computation in parallel
+        instead of creating a new one. You must still specify *njobs* as the
+        number of processes in the pool. Use this to avoid spawning processes
+        on each call to this functions, which can have significant overhead.
 
     Returns:
 
@@ -445,24 +482,15 @@ def gy(lon, lat, height, model, dens=None, ratio=RATIO_G, engine='default',
         The calculated field in mGal
 
     """
-    result = _check_input(lon, lat, height, model, ratio, njobs)
     field = 'gy'
-    if njobs == 1:
-        _dispatcher([lon, lat, height, result, model, dens, ratio, engine,
-                     field])
-    else:
-        chunks = _split_arrays(arrays=[lon, lat, height, result],
-                               extra_args=[model, dens, ratio, engine, field],
-                               nparts=njobs)
-        pool = multiprocessing.Pool(njobs)
-        result = np.hstack(pool.map(_dispatcher, chunks))
-        pool.close()
+    result = _dispatcher(field, lon, lat, height, model, dens=dens,
+                         ratio=ratio, engine=engine, njobs=njobs, pool=pool)
     result *= SI2MGAL*G
     return result
 
 
 def gz(lon, lat, height, model, dens=None, ratio=RATIO_G, engine='default',
-       njobs=1):
+       njobs=1, pool=None):
     """
     Calculate the radial component of the gravitational attraction.
 
@@ -499,6 +527,11 @@ def gz(lon, lat, height, model, dens=None, ratio=RATIO_G, engine='default',
     * njobs : int
         Split the computation into *njobs* parts and run it in parallel using
         ``multiprocessing``. If ``njobs=1`` will run the computation in serial.
+    * pool : None or multiprocessing.Pool object
+        If not None, will use this pool to run the computation in parallel
+        instead of creating a new one. You must still specify *njobs* as the
+        number of processes in the pool. Use this to avoid spawning processes
+        on each call to this functions, which can have significant overhead.
 
     Returns:
 
@@ -506,24 +539,15 @@ def gz(lon, lat, height, model, dens=None, ratio=RATIO_G, engine='default',
         The calculated field in mGal
 
     """
-    result = _check_input(lon, lat, height, model, ratio, njobs)
     field = 'gz'
-    if njobs == 1:
-        _dispatcher([lon, lat, height, result, model, dens, ratio, engine,
-                     field])
-    else:
-        chunks = _split_arrays(arrays=[lon, lat, height, result],
-                               extra_args=[model, dens, ratio, engine, field],
-                               nparts=njobs)
-        pool = multiprocessing.Pool(njobs)
-        result = np.hstack(pool.map(_dispatcher, chunks))
-        pool.close()
+    result = _dispatcher(field, lon, lat, height, model, dens=dens,
+                         ratio=ratio, engine=engine, njobs=njobs, pool=pool)
     result *= SI2MGAL*G
     return result
 
 
 def gxx(lon, lat, height, model, dens=None, ratio=RATIO_GG, engine='default',
-        njobs=1):
+        njobs=1, pool=None):
     """
     Calculate the xx component of the gravity gradient tensor.
 
@@ -555,6 +579,11 @@ def gxx(lon, lat, height, model, dens=None, ratio=RATIO_GG, engine='default',
     * njobs : int
         Split the computation into *njobs* parts and run it in parallel using
         ``multiprocessing``. If ``njobs=1`` will run the computation in serial.
+    * pool : None or multiprocessing.Pool object
+        If not None, will use this pool to run the computation in parallel
+        instead of creating a new one. You must still specify *njobs* as the
+        number of processes in the pool. Use this to avoid spawning processes
+        on each call to this functions, which can have significant overhead.
 
     Returns:
 
@@ -562,24 +591,15 @@ def gxx(lon, lat, height, model, dens=None, ratio=RATIO_GG, engine='default',
         The calculated field in Eotvos
 
     """
-    result = _check_input(lon, lat, height, model, ratio, njobs)
     field = 'gxx'
-    if njobs == 1:
-        _dispatcher([lon, lat, height, result, model, dens, ratio, engine,
-                     field])
-    else:
-        chunks = _split_arrays(arrays=[lon, lat, height, result],
-                               extra_args=[model, dens, ratio, engine, field],
-                               nparts=njobs)
-        pool = multiprocessing.Pool(njobs)
-        result = np.hstack(pool.map(_dispatcher, chunks))
-        pool.close()
+    result = _dispatcher(field, lon, lat, height, model, dens=dens,
+                         ratio=ratio, engine=engine, njobs=njobs, pool=pool)
     result *= SI2EOTVOS*G
     return result
 
 
 def gxy(lon, lat, height, model, dens=None, ratio=RATIO_GG, engine='default',
-        njobs=1):
+        njobs=1, pool=None):
     """
     Calculate the xy component of the gravity gradient tensor.
 
@@ -611,6 +631,11 @@ def gxy(lon, lat, height, model, dens=None, ratio=RATIO_GG, engine='default',
     * njobs : int
         Split the computation into *njobs* parts and run it in parallel using
         ``multiprocessing``. If ``njobs=1`` will run the computation in serial.
+    * pool : None or multiprocessing.Pool object
+        If not None, will use this pool to run the computation in parallel
+        instead of creating a new one. You must still specify *njobs* as the
+        number of processes in the pool. Use this to avoid spawning processes
+        on each call to this functions, which can have significant overhead.
 
     Returns:
 
@@ -618,24 +643,15 @@ def gxy(lon, lat, height, model, dens=None, ratio=RATIO_GG, engine='default',
         The calculated field in Eotvos
 
     """
-    result = _check_input(lon, lat, height, model, ratio, njobs)
     field = 'gxy'
-    if njobs == 1:
-        _dispatcher([lon, lat, height, result, model, dens, ratio, engine,
-                     field])
-    else:
-        chunks = _split_arrays(arrays=[lon, lat, height, result],
-                               extra_args=[model, dens, ratio, engine, field],
-                               nparts=njobs)
-        pool = multiprocessing.Pool(njobs)
-        result = np.hstack(pool.map(_dispatcher, chunks))
-        pool.close()
+    result = _dispatcher(field, lon, lat, height, model, dens=dens,
+                         ratio=ratio, engine=engine, njobs=njobs, pool=pool)
     result *= SI2EOTVOS*G
     return result
 
 
 def gxz(lon, lat, height, model, dens=None, ratio=RATIO_GG, engine='default',
-        njobs=1):
+        njobs=1, pool=None):
     """
     Calculate the xz component of the gravity gradient tensor.
 
@@ -667,6 +683,11 @@ def gxz(lon, lat, height, model, dens=None, ratio=RATIO_GG, engine='default',
     * njobs : int
         Split the computation into *njobs* parts and run it in parallel using
         ``multiprocessing``. If ``njobs=1`` will run the computation in serial.
+    * pool : None or multiprocessing.Pool object
+        If not None, will use this pool to run the computation in parallel
+        instead of creating a new one. You must still specify *njobs* as the
+        number of processes in the pool. Use this to avoid spawning processes
+        on each call to this functions, which can have significant overhead.
 
     Returns:
 
@@ -674,24 +695,15 @@ def gxz(lon, lat, height, model, dens=None, ratio=RATIO_GG, engine='default',
         The calculated field in Eotvos
 
     """
-    result = _check_input(lon, lat, height, model, ratio, njobs)
     field = 'gxz'
-    if njobs == 1:
-        _dispatcher([lon, lat, height, result, model, dens, ratio, engine,
-                     field])
-    else:
-        chunks = _split_arrays(arrays=[lon, lat, height, result],
-                               extra_args=[model, dens, ratio, engine, field],
-                               nparts=njobs)
-        pool = multiprocessing.Pool(njobs)
-        result = np.hstack(pool.map(_dispatcher, chunks))
-        pool.close()
+    result = _dispatcher(field, lon, lat, height, model, dens=dens,
+                         ratio=ratio, engine=engine, njobs=njobs, pool=pool)
     result *= SI2EOTVOS*G
     return result
 
 
 def gyy(lon, lat, height, model, dens=None, ratio=RATIO_GG, engine='default',
-        njobs=1):
+        njobs=1, pool=None):
     """
     Calculate the yy component of the gravity gradient tensor.
 
@@ -723,6 +735,11 @@ def gyy(lon, lat, height, model, dens=None, ratio=RATIO_GG, engine='default',
     * njobs : int
         Split the computation into *njobs* parts and run it in parallel using
         ``multiprocessing``. If ``njobs=1`` will run the computation in serial.
+    * pool : None or multiprocessing.Pool object
+        If not None, will use this pool to run the computation in parallel
+        instead of creating a new one. You must still specify *njobs* as the
+        number of processes in the pool. Use this to avoid spawning processes
+        on each call to this functions, which can have significant overhead.
 
     Returns:
 
@@ -730,24 +747,15 @@ def gyy(lon, lat, height, model, dens=None, ratio=RATIO_GG, engine='default',
         The calculated field in Eotvos
 
     """
-    result = _check_input(lon, lat, height, model, ratio, njobs)
     field = 'gyy'
-    if njobs == 1:
-        _dispatcher([lon, lat, height, result, model, dens, ratio, engine,
-                     field])
-    else:
-        chunks = _split_arrays(arrays=[lon, lat, height, result],
-                               extra_args=[model, dens, ratio, engine, field],
-                               nparts=njobs)
-        pool = multiprocessing.Pool(njobs)
-        result = np.hstack(pool.map(_dispatcher, chunks))
-        pool.close()
+    result = _dispatcher(field, lon, lat, height, model, dens=dens,
+                         ratio=ratio, engine=engine, njobs=njobs, pool=pool)
     result *= SI2EOTVOS*G
     return result
 
 
 def gyz(lon, lat, height, model, dens=None, ratio=RATIO_GG, engine='default',
-        njobs=1):
+        njobs=1, pool=None):
     """
     Calculate the yz component of the gravity gradient tensor.
 
@@ -779,6 +787,11 @@ def gyz(lon, lat, height, model, dens=None, ratio=RATIO_GG, engine='default',
     * njobs : int
         Split the computation into *njobs* parts and run it in parallel using
         ``multiprocessing``. If ``njobs=1`` will run the computation in serial.
+    * pool : None or multiprocessing.Pool object
+        If not None, will use this pool to run the computation in parallel
+        instead of creating a new one. You must still specify *njobs* as the
+        number of processes in the pool. Use this to avoid spawning processes
+        on each call to this functions, which can have significant overhead.
 
     Returns:
 
@@ -786,24 +799,15 @@ def gyz(lon, lat, height, model, dens=None, ratio=RATIO_GG, engine='default',
         The calculated field in Eotvos
 
     """
-    result = _check_input(lon, lat, height, model, ratio, njobs)
     field = 'gyz'
-    if njobs == 1:
-        _dispatcher([lon, lat, height, result, model, dens, ratio, engine,
-                     field])
-    else:
-        chunks = _split_arrays(arrays=[lon, lat, height, result],
-                               extra_args=[model, dens, ratio, engine, field],
-                               nparts=njobs)
-        pool = multiprocessing.Pool(njobs)
-        result = np.hstack(pool.map(_dispatcher, chunks))
-        pool.close()
+    result = _dispatcher(field, lon, lat, height, model, dens=dens,
+                         ratio=ratio, engine=engine, njobs=njobs, pool=pool)
     result *= SI2EOTVOS*G
     return result
 
 
 def gzz(lon, lat, height, model, dens=None, ratio=RATIO_GG, engine='default',
-        njobs=1):
+        njobs=1, pool=None):
     """
     Calculate the zz component of the gravity gradient tensor.
 
@@ -835,6 +839,11 @@ def gzz(lon, lat, height, model, dens=None, ratio=RATIO_GG, engine='default',
     * njobs : int
         Split the computation into *njobs* parts and run it in parallel using
         ``multiprocessing``. If ``njobs=1`` will run the computation in serial.
+    * pool : None or multiprocessing.Pool object
+        If not None, will use this pool to run the computation in parallel
+        instead of creating a new one. You must still specify *njobs* as the
+        number of processes in the pool. Use this to avoid spawning processes
+        on each call to this functions, which can have significant overhead.
 
     Returns:
 
@@ -842,17 +851,8 @@ def gzz(lon, lat, height, model, dens=None, ratio=RATIO_GG, engine='default',
         The calculated field in Eotvos
 
     """
-    result = _check_input(lon, lat, height, model, ratio, njobs)
     field = 'gzz'
-    if njobs == 1:
-        _dispatcher([lon, lat, height, result, model, dens, ratio, engine,
-                     field])
-    else:
-        chunks = _split_arrays(arrays=[lon, lat, height, result],
-                               extra_args=[model, dens, ratio, engine, field],
-                               nparts=njobs)
-        pool = multiprocessing.Pool(njobs)
-        result = np.hstack(pool.map(_dispatcher, chunks))
-        pool.close()
+    result = _dispatcher(field, lon, lat, height, model, dens=dens,
+                         ratio=ratio, engine=engine, njobs=njobs, pool=pool)
     result *= SI2EOTVOS*G
     return result